createNewDateList()をanalysisGeneral.tsに移動 #1520
